Try replacing 'specification' with document/profile
We have the sentence "This document is not a specification, but a profile.",
but then use the word "specification" a lot to describe the document.

The required OIDF boilerplate text also refers to it as a specifications,
and I'm not sure we can change that... so maybe we should do something
about the "This document is not a specification" sentence instead.