Tweak 'This document is not a specification' text
We have the sentence "This document is not a specification, but a
profile.", but then use the word "specification" a lot to describe the
document.

The required OIDF "Notices" boilerplate text also refers to it as a
specifications, so reword the 'not a specification' text instead so
we at least don't immediately contradict ourselves.