Tweak 'This document is not a specification' text (#311)
4 approvals, open over a week, no unresolved comments - merging!

* Tweak 'This document is not a specification' text

We have the sentence "This document is not a specification, but a
profile.", but then use the word "specification" a lot to describe the
document.

The required OIDF "Notices" boilerplate text also refers to it as a
specifications, so reword the 'not a specification' text instead so
we at least don't immediately contradict ourselves.

* Apply Kristina's suggestion

Co-authored-by: Kristina <52878547+Sakurann@users.noreply.github.com>

---------

Co-authored-by: Kristina <52878547+Sakurann@users.noreply.github.com>